Only run the system partition resizer if a rootfs is included
This allows us to break out of a chicken-and-egg problem where the
update that fixes corruption on SD writes cannot be installed because
the files are corrupted while installing. Technically it does nothing
to avoid corruption, but installing only a kernel writes much less
data, so it is far more likely to succeed without corrupting.